Refactor shim client to another file
