+++ pkg/osvscanner/vulnerability_result.go
@@ -62,7 +62,7 @@ func groupResponseBySource(r *output.Reporter, query osv.BatchedQuery, resp *osv
 				r.PrintError("Failed to run govulncheck on " + si.Path)
 			} else {
 				for _, v := range res.Vulns {
-					r.PrintText(fmt.Sprintf("%+v -- %v\n", v, v.IsCalled()))
+					r.PrintText(fmt.Sprintf("%+v -- \n", v))
 				}
 			}
 		}
